Add data persistance

+++ BusShuttleProject/BusShuttle/FileSaver.cs
@@ -8,7 +8,10 @@ public class FileSaver
   public FileSaver(string fileName)
   {
     this.fileName = fileName;
-    File.Create(this.fileName).Close();
+    if (!File.Exists(this.fileName))
+    {
+      File.Create(this.fileName).Close();
+    }
   }
 
   public void AppendLine(string line)
